+++ apps/web/app/composables/index.ts
@@ -103,4 +103,5 @@ export * from './useCustomAssets';
 export * from './useItemDataTable';
 export * from './useBlockClasses';
 export * from './useCallisto';
-export * from './useUtilityBarSection';
+export * from './useUtilityBarActions';
+export * from './useUtilityBarForm';
