--- apps/web/app/composables/useUtilityBarSection/index.ts
@@ -1,2 +1,0 @@
-export * from './types';
-export * from './useUtilityBarSection';
